+++ src/com/intellij/idea/plugin/hybris/impex/lang/refactoring/ImpExPsiElementManipulator.kt
@@ -1,10 +1,10 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
- * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -16,14 +16,14 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.impex.rename.manipulator
+package com.intellij.idea.plugin.hybris.impex.lang.refactoring
 
 import com.intellij.idea.plugin.hybris.impex.psi.util.setName
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.AbstractElementManipulator
 import com.intellij.psi.PsiElement
 
-class ImpexMacrosManipulator : AbstractElementManipulator<PsiElement>() {
+class ImpExPsiElementManipulator : AbstractElementManipulator<PsiElement>() {
 
     override fun handleContentChange(element: PsiElement, range: TextRange, newName: String): PsiElement = setName(element, newName)