+++ src/com/intellij/idea/plugin/hybris/impex/lang/refactoring/ImpExRenamePsiElementProcessor.kt
@@ -1,10 +1,10 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
- * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -16,17 +16,15 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.impex.rename.processor
+package com.intellij.idea.plugin.hybris.impex.lang.refactoring
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
-import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroNameDec
-import com.intellij.idea.plugin.hybris.impex.psi.ImpexMacroUsageDec
-import com.intellij.idea.plugin.hybris.impex.psi.ImpexPsiNamedElement
-import com.intellij.idea.plugin.hybris.impex.psi.ImpexTypes
-import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexMacroReference
+import com.intellij.idea.plugin.hybris.impex.psi.*
 import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiReference
 import com.intellij.psi.search.SearchScope
+import com.intellij.psi.util.PsiElementFilter
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.elementType
 import com.intellij.refactoring.listeners.RefactoringElementListener
@@ -34,14 +32,18 @@ import com.intellij.refactoring.rename.RenamePsiElementProcessor
 import com.intellij.refactoring.rename.UnresolvableCollisionUsageInfo
 import com.intellij.usageView.UsageInfo
 
-class ImpexMacrosRenameProcessor : RenamePsiElementProcessor() {
+class ImpExRenamePsiElementProcessor : RenamePsiElementProcessor() {
 
     override fun canProcessElement(element: PsiElement) = element is ImpexMacroNameDec
+        || element is ImpexDocumentIdDec
+        || element is ImpexDocumentIdUsage
         || (element is ImpexMacroUsageDec && !element.text.startsWith(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX))
 
-    override fun findReferences(element: PsiElement, searchScope: SearchScope, searchInCommentsAndStrings: Boolean) = findElements(element, element.text)
-        .map { ImpexMacroReference(it) }
-        .toMutableList()
+    override fun findReferences(element: PsiElement, searchScope: SearchScope, searchInCommentsAndStrings: Boolean): Collection<PsiReference> {
+        return findElements(element, element.text)
+            .mapNotNull { it.reference }
+            .toMutableList()
+    }
 
     override fun renameElement(element: PsiElement, newName: String, usages: Array<out UsageInfo>, listener: RefactoringElementListener?) {
         with(element as ImpexPsiNamedElement) {
@@ -85,12 +87,31 @@ class ImpexMacrosRenameProcessor : RenamePsiElementProcessor() {
             }
     }
 
-    private fun findElements(element: PsiElement, newText: String): Array<PsiElement> = element.containingFile
+    private fun findElements(renameElement: PsiElement, newText: String): Array<PsiElement> = renameElement.containingFile
         ?.let { file ->
-            PsiTreeUtil.collectElements(file) {
-                (it is ImpexMacroNameDec && it.textMatches(newText))
-                    || (it is ImpexMacroUsageDec && it.reference?.resolve()?.text == newText)
+            val filter = when (renameElement) {
+                is ImpexMacroNameDec,
+                is ImpexMacroUsageDec -> PsiElementFilter { element ->
+                    if (element is ImpexMacroNameDec || element is ImpexMacroUsageDec) {
+                        element.text == newText
+                    } else {
+                        false
+                    }
+                }
+
+                is ImpexDocumentIdDec,
+                is ImpexDocumentIdUsage -> PsiElementFilter { element ->
+                    if (element is ImpexDocumentIdDec || element is ImpexDocumentIdUsage) {
+                        element.text == newText
+                    } else {
+                        false
+                    }
+                }
+
+                else -> return@let null
             }
+
+            PsiTreeUtil.collectElements(file, filter)
         }
         ?: emptyArray()
 }