+++ src/com/intellij/idea/plugin/hybris/impex/lang/refactoring/ImpExVetoRenameCondition.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -15,13 +15,14 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.intellij.idea.plugin.hybris.impex.rename
+
+package com.intellij.idea.plugin.hybris.impex.lang.refactoring
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.openapi.util.Condition
 import com.intellij.psi.PsiElement
 
-class ImpexVetoRenameCondition : Condition<PsiElement> {
+class ImpExVetoRenameCondition : Condition<PsiElement> {
 
     override fun value(t: PsiElement?) = t?.text?.startsWith(HybrisConstants.IMPEX_CONFIG_COMPLETE_PREFIX)
         ?: false