+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Added Item values reference resolution within ComposedType attribute columns [#1361](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1361)
 - Resolve macro usages in the header parameter for attribute sub-type [#1362](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1362)
 - Added code completion and reference resolution for `&DocId` usages [#1364](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1364)
+- Added rename capabilities for Document ID `&DocId` elements [#1365](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1365)
 
 ### `Terminal` integration
 - Added support of the `Terminal` plugin and set default directory to the project root for new Terminal Windows [#1342](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1342)